Mapeos

+++ src/main/java/clasesDAO/AlumnoDAOHibernateJPA.java
@@ -18,14 +18,14 @@ public class AlumnoDAOHibernateJPA extends GenericDAOHibernateJPA<Alumno> implem
 	}
 	
 	@Override
-	public List<Comentario> obtenerComentarios(int id) {
+	public List<Comentario> obtenerComentarios(long id) {
 		Query q = EMF.getEMF().createEntityManager().createQuery("Select * from Alumno a INNER JOIN Comentario c on a.id=c.idP");
 		List<Comentario> resultado = (List<Comentario>) q.getResultList();
 		return resultado;
 	}
 
 	@Override
-	public List<String> obtenerIntereses(int id) { //Ver como guardar intereses
+	public List<String> obtenerIntereses(long id) { //Ver como guardar intereses
 		Query q = EMF.getEMF().createEntityManager().createQuery("Select * from Alumno a INNER JOIN Interes i on a.id=i.idP");
 		List<String> resultado = (List<String>) q.getResultList();
 		return resultado;
