+++ src/main/java/clasesDAO/CarteleraDAOHibernateJPA.java
@@ -18,14 +18,14 @@ public class CarteleraDAOHibernateJPA extends GenericDAOHibernateJPA<Cartelera> 
 	}
 	
 	@Override
-	public List<Publicacion> obtenerPublicaciones(int id) {
+	public List<Publicacion> obtenerPublicaciones(long id) {
 		Query q = EMF.getEMF().createEntityManager().createQuery("Select * from Cartelera c INNER JOIN Publicacion p on c.id=p.idC");
 		List<Publicacion> resultado = (List<Publicacion>) q.getResultList();
 		return resultado;
 	}
 
 	@Override
-	public List<Alumno> obtenerAlumnosInteresados(int id) {
+	public List<Alumno> obtenerAlumnosInteresados(long id) {
 		Query q = EMF.getEMF().createEntityManager().createQuery("Select * from Cartelera c INNER JOIN Intereses i on c.id=i.idC");
 		List<Alumno> resultado = (List<Alumno>) q.getResultList();
 		return resultado;
